feat: prevent deleting transaction if status = COMPLETED

+++ server/src/main/java/com/pdmstudents/hpesystem/service/OrderService.java
@@ -135,13 +135,9 @@ public class OrderService {
     repo.findById(id)
         .ifPresentOrElse(
             order -> {
-              order
-                  .getOrderItems()
-                  .forEach(
-                      item -> {
-                        var part = item.getPart();
-                        part.setStock(part.getStock() + item.getQuantity());
-                      });
+              if (order.getStatus() == OrderStatus.COMPLETED) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+              }
               repo.delete(order);
             },
             () -> {
